refining regex match in cfdb:property to determine whether $value is a JSON array or not
cfdb:property serializes a key/value pair as a json property construct "$key":"$value". In order to determine whether to put quotation marks around $value it uses
a) the Xquery type of $value (xs:integer = no quotation marks) OR
b) a regex to check if the string in $value is a already serialized json array (in case it would start/end with square brackets and contain a colon)

WATCHME This is a quick workaround because of course there might be string literals that start/end with square brackets AND contain a colon - in this case they would be interpreted as an array.